src/kawaz/core/google/calendar/requirements.txt
httplib2
git+https://github.com/enorvelle/GoogleApiPython3x.git#egg=google_api_python_client
tolerance
django-appconf
